Day 16 finally.
This one was tough for me - needed to just try a few things for part1.
For part 2, I gave up, and decided to store all the paths, and find the
best two. It started really slow and memory intensive, but pulling out
the not-requried bits, and terminating the search loop properly for the
two paths made it pretty acceptable in time and space.